Try to restore `yarn mocha` and move browser tests to `/browser/`
